scraper/requirements.txt
linkedin_jobs_scraper==5.0.2
python-dotenv==1.1.1
selenium==4.33.0
